// src/app/shared/models/ph.interface.ts
export interface PH {
    al:                             number;
    del:                            number;
    rechazados:                     number;
    reses:                          number;
    resesConPHYSegregados:          number[];
    segregados:                     number[];
    tropa:                          number;
}